Add wrapper for CollectionList in profile
