+++ src/components/profile-collection/CollectionList.jsx
@@ -1,7 +1,7 @@
 import { IonInfiniteScroll, IonInfiniteScrollContent, IonList } from "@ionic/react";
 import { useHistory } from "react-router";
 import InfiniteScroll from "../infinite-scroll/InfiniteScroll";
-import ProfileCollection from "./ProfileCollection";
+import CollectionCard from "./CollectionCard";
 
 const CollectionList = (props) => {
   const history = useHistory();
@@ -10,7 +10,7 @@ const CollectionList = (props) => {
   return (
     <IonList className="profile-collection--list">
       {collections.map((collection, index) => (
-        <ProfileCollection collection={collection} key={index} onClick={() => history.push(`/collections/${collection.collectionId}`)} />
+        <CollectionCard collection={collection} key={index} onClick={() => history.push(`/collections/${collection.collectionId}`)} />
       ))}
       <InfiniteScroll onScrollEnd={fetchNextPage} listEnded={listEnded} />
     </IonList>
